// frontend/src/types/bindings/Session.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Team } from "./Team";

/**
 * Represents a session in the Bible Challenge.
 */
export type Session = { teams: Array<Team>, buzz_lock: boolean, dark_mode: boolean, timer_enabled: boolean, current_page: string, created_at: string, last_modified: string, };